260 fix api view so that it doesn't access database at import time
