+++ vitrina/api/urls.py
@@ -2,10 +2,9 @@ from django.urls import path, include
 from django.views.generic import TemplateView
 from rest_framework.routers import DefaultRouter
 
-# from vitrina.api.services import get_partner_schema_view
 from vitrina.api.views import CatalogViewSet, DatasetViewSet, CategoryViewSet, LicenceViewSet, \
     DatasetDistributionViewSet, DatasetStructureViewSet, InternalDatasetViewSet, InternalDatasetDistributionViewSet, \
-    InternalDatasetStructureViewSet
+    InternalDatasetStructureViewSet, PartnerApiView
 
 router = DefaultRouter(trailing_slash=False)
 router.register(r'catalogs', CatalogViewSet, basename="api-catalog")
@@ -13,7 +12,7 @@ router.register(r'categories', CategoryViewSet, basename="api-category")
 router.register(r'licences', LicenceViewSet, basename="api-licence")
 
 urlpatterns = [
-    # path('partner/api/1/', get_partner_schema_view().with_ui('redoc'), name='partner-api'),
+    path('partner/api/1/', PartnerApiView.with_ui('redoc'), name='partner-api'),
     path('partner/api/1/', include(router.urls)),
 
     # dataset api urls
